I action result

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -21,9 +21,10 @@ namespace HotelFinder.API.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        public List<Hotel> Get()
+        public IActionResult Get()
         {
-            return _hotelService.GetAllHotels();
+            var hotels = _hotelService.GetAllHotels();
+            return Ok(hotels); // 200 + data
         }
 
 
@@ -33,9 +34,15 @@ namespace HotelFinder.API.Controllers
         /// <param name="id"></param>
         /// <returns></returns>
         [HttpGet("{id}")]
-        public Hotel Get(int id)
+        public IActionResult Get(int id)
         {
-            return _hotelService.GetHotelById(id);
+            var hotel = _hotelService.GetHotelById(id);
+
+            if (hotel!= null)
+            {
+                return Ok(hotel); //200 + data
+            }
+            return NotFound(); //404
         }
 
         /// <summary>
@@ -44,9 +51,14 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPost]
-        public Hotel Post([FromBody] Hotel hotel)
+        public IActionResult Post([FromBody] Hotel hotel)
         {
-            return _hotelService.CreateHotel(hotel);
+            // if (ModelState.IsValid) "[ApiController]" validation işlemlerini kendi otamatik kontrol ediyor
+            //{                                        data geçerli değilse bu actiona girmiyor bile
+            var createdHotel = _hotelService.CreateHotel(hotel);
+            return CreatedAtAction("Get", new { id = createdHotel.Id }, createdHotel); //201 + data
+            //}
+            //return BadRequest(ModelState); //404 + validation errors
         }
 
         /// <summary>
@@ -55,9 +67,14 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPut]
-        public Hotel Put([FromBody] Hotel hotel)
+        public IActionResult Put([FromBody] Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            if (_hotelService.GetHotelById(hotel.Id)!=null)
+            {
+                return Ok(_hotelService.UpdateHotel(hotel)); //200 + data
+            }
+            
+            return NotFound();
         }
 
 
@@ -66,9 +83,16 @@ namespace HotelFinder.API.Controllers
         /// </summary>
         /// <param name="id"></param>
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-            _hotelService.DeleteHotel(id);
+            if (_hotelService.GetHotelById(id) != null)
+            {
+                _hotelService.DeleteHotel(id);
+                return Ok(); //200
+            }
+
+            return NotFound();
+            
         }
     }
 }
